fix: rename source files in app/ which conflict with 3rdparty/qmarkdowntextedit
