+++ app/cutemarkedmainwindow.cpp
@@ -14,8 +14,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "mainwindow.h"
-#include "ui_mainwindow.h"
+#include "cutemarkedmainwindow.h"
+#include "ui_cutemarkedmainwindow.h"
 
 #include <QClipboard>
 #include <QDesktopServices>
